requirements.txt
# requirements.txt
mcp
anthropic
fastapi
uvicorn[standard]
pydantic
aiohttp
websockets
python-dotenv